requirements.txt
asgiref==3.5.0
astroid==2.11.3
black==22.3.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.2
cryptography==36.0.2
cv==1.0.0
defusedxml==0.7.1
dill==0.3.4
Django==3.0
django-allauth==0.50.0
django-cors-headers==3.11.0
django-rest-auth==0.9.5
djangorestframework==3.13.1
gunicorn==20.1.0
idna==3.3
isort==5.10.1
lazy-object-proxy==1.7.1
mccabe==0.7.0
mypy-extensions==0.4.3
numpy==1.22.3
oauthlib==3.2.0
opencv-contrib-python==4.5.5.64
opencv-python==4.5.5.64
pathspec==0.9.0
Pillow==9.1.0
platformdirs==2.5.2
psycopg2-binary==2.9.3
pycparser==2.21
PyJWT==2.3.0
pylint==2.13.7
python3-openid==3.2.0
pytz==2022.1
requests==2.27.1
requests-oauthlib==1.3.1
six==1.16.0
sqlparse==0.4.2
stripe==2.72.0
tomli==2.0.1
typing_extensions==4.2.0
urllib3==1.26.9
wrapt==1.11.2
